requirements.txt
fastapi==0.111.0
pydantic==2.5.3
python-dotenv==1.0.1
uvicorn==0.30.1
langchain==0.2.1
langchain-core==0.2.3 
langchain-groq==0.1.4  
